Fix TJSAMP_420 type mismatch and Swift 6 Sendable warnings
- Cast TJSAMP_420.rawValue to Int32 to match tjCompress2 parameter type
- Use nonisolated(unsafe) for ChannelHandlerContext captures in
  @Sendable closures (safe because execution is on the event loop)
- Copy var buffers to let bindings before capture in Sendable closures
- Move guard let self out of eventLoop.execute in SSE handler

+++ Sources/XToolSupport/Preview/JPEGCompressor.swift
@@ -134,7 +134,7 @@ enum JPEGCompressor {
                 pixelFormat.rawValue,
                 &jpegBuf,
                 &jpegSize,
-                TJSAMP_420.rawValue,
+                Int32(TJSAMP_420.rawValue),
                 Int32(quality),
                 Int32(TJFLAG_FASTDCT)
             )
